chore: create SearchParamsRaw
